Move base interfaces
