+++ src/main/java/fr/epicanard/globalmarketchest/gui/shops/interfaces/ConfigShop.java
@@ -1,7 +1,7 @@
 package fr.epicanard.globalmarketchest.gui.shops.interfaces;
 
 import fr.epicanard.globalmarketchest.gui.InventoryGUI;
-import fr.epicanard.globalmarketchest.gui.shops.ShopInterface;
+import fr.epicanard.globalmarketchest.gui.shops.baseinterfaces.ShopInterface;
 
 public class ConfigShop extends ShopInterface {
