+++ src/main/java/fr/epicanard/globalmarketchest/gui/shops/interfaces/ShopCreationLink.java
@@ -10,7 +10,7 @@ import fr.epicanard.globalmarketchest.exceptions.ShopAlreadyExistException;
 import fr.epicanard.globalmarketchest.gui.InventoryGUI;
 import fr.epicanard.globalmarketchest.gui.TransactionKey;
 import fr.epicanard.globalmarketchest.gui.paginator.Paginator;
-import fr.epicanard.globalmarketchest.gui.shops.ShopCreationInterface;
+import fr.epicanard.globalmarketchest.gui.shops.baseinterfaces.ShopCreationInterface;
 import fr.epicanard.globalmarketchest.gui.actions.ChatInput;
 import fr.epicanard.globalmarketchest.gui.actions.LeaveShop;
 import fr.epicanard.globalmarketchest.gui.actions.PreviousInterface;
